+++ app/src/main/java/org/sopt/dosopttemplate/data/dto/request/RequestUserDataDto.kt
@@ -4,8 +4,7 @@ import com.google.gson.annotations.SerializedName
 import kotlinx.serialization.Serializable
 
 @Serializable
-
-data class RequestInquiryDto (
+data class RequestUserDataDto (
     @SerializedName("id")
     val id:Int,
 )
