+++ app/src/main/java/org/sopt/dosopttemplate/data/dto/response/ResponseUserDataDto.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 
-data class ResponseInquiryDto(
+data class ResponseUserDataDto(
     @SerializedName("username")
     val username: String,
